+++ packages/core/src/components/Table/useGroupedRowSelector.ts
@@ -1,19 +1,22 @@
-import { useCallback, useState } from 'react';
+import { Dispatch, SetStateAction, useCallback } from 'react';
 
 type Result = {
     selectedIds: (string | number)[];
     toggleIds: (ids: any[]) => void;
 };
-const useGroupedRowSelector = (): Result => {
-    const [selectedIds, setSelectedIds] = useState<(string | number)[]>([]);
+type Params = {
+    selectedIds?: (string | number)[];
+    setSelectedIds: Dispatch<SetStateAction<(string | number)[]>>;
+};
 
+const useGroupedRowSelector = ({ selectedIds, setSelectedIds }: Params): Result => {
     const toggleIds = useCallback(
         (ids: number[] | string[]) =>
             setSelectedIds(sIds =>
                 // @ts-ignore
                 ids.reduce((acc, cur) => (acc.indexOf(cur) === -1 ? [...acc, cur] : acc.filter(id => id !== cur)), [...sIds])
             ),
-        [setSelectedIds]
+        []
     );
 
     return { selectedIds, toggleIds };