+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.3](https://github.com/medly/medly-components/compare/@medly-components/icons@1.3.2...@medly-components/icons@1.3.3) (2020-09-14)
+
+**Note:** Version bump only for package @medly-components/icons
+
+
+
+
+
 ## [1.3.2](https://github.com/medly/medly-components/compare/@medly-components/icons@1.3.1...@medly-components/icons@1.3.2) (2020-09-10)
 
 **Note:** Version bump only for package @medly-components/icons